// types/elevation.ts
// elevation.ts
export interface Elevation {
    id: string;
    plotId: string;
    jobCode: string;
    plotNumber: string;
    houseType: string;
    houseNo: string | null;
    street: string | null;
    town: string | null;
    postcode: string | null;
    mpan: string;
    panel: string;
    panelKwp: number;
    mcsCode: string;
    orientation: string;
    columns: number;
    rows: number;
    phase: number;
    totalStrings: number;
    string1: string;
    string2: string;
    string3: string;
    string4: string;
    inverter: string;
    trackerStringNo: string;
    inverterHybrid: string | null;
    typeTestNo: string;
    ratedOutputPower: number | null;
    battery: string;
    mountingKit: string;
    tileType: string;
    roofIncline: number;
    variationFromSouth: string;
    kwhPerKwp: number | null;
    inAboveRoof: number | null;
    overshadingFactor: number | null;
    noPanels: number;
    arrayM2: number | null;
    kwp: number;
    kwh: number | null;
    co2Equivalent: number | null;
    netKwp: number;
    meterMake: string | null;
    meterModel: string | null;
    totalCost: number;
    totalPrice: string;
    givenEnergy: string;
    instanceId: string;
}